Modified the draw property and added an innerFrame property to the sprites.
The draw property is now an array of functions. This is to make use of
the new innerFrame property of the sprites. The innerFrame property is
a function that takes in a number (the current frame) and returns the
index of the draw function to be used. This allows the user to choose
which draw function to use at any given frame (given they can make an
algorithm for it), while still keeping the keyframe-tweener independent.