Fixed the bird model, worked on the animations. Balloon is working, birds need some work.
Also revamped the InnerFrame property. It is now a number that
signifies how often to change the model. The new currentInner property
tracks the current inner frame for the tweener so it can properly cycle
through them. The old function support has been made an optional
feature that the tweener will look for.